requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
sentence-transformers
requests
python-dotenv